Cuda test

+++ src/CPPAlgos/BlockLRACPPAlgo.cpp
@@ -16,9 +16,9 @@ namespace AMMBench {
   torch::Tensor AMMBench::BlockLRACPPAlgo::amm(torch::Tensor A, torch::Tensor B, uint64_t blockSize) {
   
   // Input size and block size
-  uint64_t m = A.size(0);
-  uint64_t k = A.size(1);
-  uint64_t n = B.size(1);
+  int64_t m = A.size(0);
+  int64_t k = A.size(1);
+  int64_t n = B.size(1);
 
   assert(m % blockSize == 0);
   assert(k % blockSize == 0);
@@ -29,7 +29,7 @@ namespace AMMBench {
 
   for (uint64_t I=0; I<m/blockSize; ++I){
     for (uint64_t J=0; J<n/blockSize; ++J){
-      torch::Tensor subFinalLRA = torch::zeros({blockSize, blockSize});
+      torch::Tensor subFinalLRA = torch::zeros({(int64_t)blockSize, (int64_t)blockSize});
       for (uint64_t K=0; K<k/blockSize; ++K){
         // get sub matrix
         torch::Tensor AIK = A.index({torch::indexing::Slice(I*blockSize, (I+1)*blockSize), torch::indexing::Slice(K*blockSize, (K+1)*blockSize)});
